K线图长按，时间

+++ FL_StockChart/FL_StockChart/FLStockChart/Category/CATextLayer+TimeTextLayer.h
@@ -22,7 +22,7 @@
  */
 + (CATextLayer *)getTextLayerWithString:(NSString *)text
                               textColor:(UIColor *)textColor
-                               fontSize:(NSInteger)fontSize
+                               fontSize:(CGFloat)fontSize
                         backgroundColor:(UIColor *)bgColor
                                   frame:(CGRect)frame;
 @end
